Fix building sock_ctrL_msg.rs against musl libc
Musl libc uses u32 for cmsg_len, cast it to usize when calling
the new CMSG_ALIGN const fn.

Bug: 216192129
Test: m USE_HOST_MUSL=true crosvm
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4521472
Commit-Queue: Colin Cross <ccross@google.com>

+++ base/src/sys/unix/sock_ctrl_msg.rs
@@ -71,7 +71,10 @@ fn CMSG_DATA(cmsg_buffer: *mut cmsghdr) -> *mut RawFd {
 // does some pointer arithmetic on cmsg_ptr.
 #[allow(clippy::cast_ptr_alignment, clippy::unnecessary_cast)]
 fn get_next_cmsg(msghdr: &msghdr, cmsg: &cmsghdr, cmsg_ptr: *mut cmsghdr) -> *mut cmsghdr {
-    let next_cmsg = (cmsg_ptr as *mut u8).wrapping_add(CMSG_ALIGN(cmsg.cmsg_len)) as *mut cmsghdr;
+    // The extra cast of cmsg_len to usize is required to build against musl libc, which uses
+    // u32 for cmsg_len.
+    let next_cmsg =
+        (cmsg_ptr as *mut u8).wrapping_add(CMSG_ALIGN(cmsg.cmsg_len as usize)) as *mut cmsghdr;
     if next_cmsg
         .wrapping_offset(1)
         .wrapping_sub(msghdr.msg_control as usize) as usize
